% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servicecontrol_objects.R
\name{AllocateQuotaRequest}
\alias{AllocateQuotaRequest}
\title{AllocateQuotaRequest Object}
\usage{
AllocateQuotaRequest(allocationMode = NULL, serviceConfigId = NULL,
  allocateOperation = NULL)
}
\arguments{
\item{allocationMode}{Allocation mode for this operation}

\item{serviceConfigId}{Specifies which version of service configuration should be used to process}

\item{allocateOperation}{Operation that describes the quota allocation}
}
\value{
AllocateQuotaRequest object
}
\description{
AllocateQuotaRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Request message for the AllocateQuota method.
}
\seealso{
Other AllocateQuotaRequest functions: \code{\link{services.allocateQuota}}
}
