% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servicecontrol_objects.R
\name{AllocateQuotaResponse}
\alias{AllocateQuotaResponse}
\title{AllocateQuotaResponse Object}
\usage{
AllocateQuotaResponse(operationId = NULL, serviceConfigId = NULL,
  allocateErrors = NULL, quotaMetrics = NULL)
}
\arguments{
\item{operationId}{The same operation_id value used in the AllocateQuotaRequest}

\item{serviceConfigId}{ID of the actual config used to process the request}

\item{allocateErrors}{Indicates the decision of the allocate}

\item{quotaMetrics}{Quota metrics to indicate the result of allocation}
}
\value{
AllocateQuotaResponse object
}
\description{
AllocateQuotaResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Response message for the AllocateQuota method.
}
