% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servicecontrol_objects.R
\name{AuditLog}
\alias{AuditLog}
\title{Google Service Control API Objects 
Google Service Control provides control plane functionality to managed services, such as logging, monitoring, and status checks.}
\usage{
AuditLog(AuditLog.response = NULL, AuditLog.request = NULL,
  AuditLog.serviceData = NULL, numResponseItems = NULL, status = NULL,
  authenticationInfo = NULL, serviceName = NULL, response = NULL,
  methodName = NULL, authorizationInfo = NULL, resourceName = NULL,
  request = NULL, requestMetadata = NULL, serviceData = NULL)
}
\arguments{
\item{AuditLog.response}{The \link{AuditLog.response} object or list of objects}

\item{AuditLog.request}{The \link{AuditLog.request} object or list of objects}

\item{AuditLog.serviceData}{The \link{AuditLog.serviceData} object or list of objects}

\item{numResponseItems}{The number of items returned from a List or Query API method,}

\item{status}{The status of the overall operation}

\item{authenticationInfo}{Authentication information}

\item{serviceName}{The name of the API service performing the operation}

\item{response}{The operation response}

\item{methodName}{The name of the service method or operation}

\item{authorizationInfo}{Authorization information}

\item{resourceName}{The resource or collection that is the target of the operation}

\item{request}{The operation request}

\item{requestMetadata}{Metadata about the operation}

\item{serviceData}{Other service-specific data about the request, response, and other}
}
\value{
AuditLog object
}
\description{
Auto-generated code by googleAuthR::gar_create_api_objects
 at 2017-03-05 20:12:13
filename: /Users/mark/dev/R/autoGoogleAPI/googleservicecontrolv1.auto/R/servicecontrol_objects.R
api_json: api_json
}
\details{
Objects for use by the functions created by googleAuthR::gar_create_api_skeleton
AuditLog Object


Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Common audit log format for Google Cloud Platform API operations.
}
\seealso{
Other AuditLog functions: \code{\link{AuditLog.request}},
  \code{\link{AuditLog.response}},
  \code{\link{AuditLog.serviceData}}
}
