% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servicecontrol_objects.R
\name{AuditLog.request}
\alias{AuditLog.request}
\title{AuditLog.request Object}
\usage{
AuditLog.request()
}
\value{
AuditLog.request object
}
\description{
AuditLog.request Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The operation request. This may not include all request parameters,such as those that are too large, privacy-sensitive, or duplicatedelsewhere in the log record.It should never include user-generated data, such as file contents.When the JSON object represented here has a proto equivalent, the protoname will be indicated in the `@type` property.
}
\seealso{
Other AuditLog functions: \code{\link{AuditLog.response}},
  \code{\link{AuditLog.serviceData}},
  \code{\link{AuditLog}}
}
