% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servicecontrol_objects.R
\name{AuditLog.serviceData}
\alias{AuditLog.serviceData}
\title{AuditLog.serviceData Object}
\usage{
AuditLog.serviceData()
}
\value{
AuditLog.serviceData object
}
\description{
AuditLog.serviceData Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Other service-specific data about the request, response, and otheractivities.
}
\seealso{
Other AuditLog functions: \code{\link{AuditLog.request}},
  \code{\link{AuditLog.response}}, \code{\link{AuditLog}}
}
