% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servicecontrol_objects.R
\name{AuthenticationInfo}
\alias{AuthenticationInfo}
\title{AuthenticationInfo Object}
\usage{
AuthenticationInfo(principalEmail = NULL, authoritySelector = NULL)
}
\arguments{
\item{principalEmail}{The email address of the authenticated user making the request}

\item{authoritySelector}{The authority selector specified by the requestor, if any}
}
\value{
AuthenticationInfo object
}
\description{
AuthenticationInfo Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Authentication information for the operation.
}
