% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servicecontrol_objects.R
\name{AuthorizationInfo}
\alias{AuthorizationInfo}
\title{AuthorizationInfo Object}
\usage{
AuthorizationInfo(resource = NULL, granted = NULL, permission = NULL)
}
\arguments{
\item{resource}{The resource being accessed, as a REST-style string}

\item{granted}{Whether or not authorization for `resource` and `permission`}

\item{permission}{The required IAM permission}
}
\value{
AuthorizationInfo object
}
\description{
AuthorizationInfo Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Authorization information for the operation.
}
