% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servicecontrol_objects.R
\name{CheckError}
\alias{CheckError}
\title{CheckError Object}
\usage{
CheckError(detail = NULL, code = NULL)
}
\arguments{
\item{detail}{Free-form text providing details on the error cause of the error}

\item{code}{The error code}
}
\value{
CheckError object
}
\description{
CheckError Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Defines the errors to be returned ingoogle.api.servicecontrol.v1.CheckResponse.check_errors.
}
