% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servicecontrol_objects.R
\name{CheckInfo}
\alias{CheckInfo}
\title{CheckInfo Object}
\usage{
CheckInfo(unusedArguments = NULL)
}
\arguments{
\item{unusedArguments}{A list of fields and label keys that are ignored by the server}
}
\value{
CheckInfo object
}
\description{
CheckInfo Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
