% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servicecontrol_objects.R
\name{CheckRequest}
\alias{CheckRequest}
\title{CheckRequest Object}
\usage{
CheckRequest(serviceConfigId = NULL, skipActivationCheck = NULL,
  operation = NULL, requestProjectSettings = NULL)
}
\arguments{
\item{serviceConfigId}{Specifies which version of service configuration should be used to process}

\item{skipActivationCheck}{Indicates if service activation check should be skipped for this request}

\item{operation}{The operation to be checked}

\item{requestProjectSettings}{Requests the project settings to be returned as part of the check response}
}
\value{
CheckRequest object
}
\description{
CheckRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Request message for the Check method.
}
\seealso{
Other CheckRequest functions: \code{\link{services.check}}
}
