% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servicecontrol_objects.R
\name{CheckResponse}
\alias{CheckResponse}
\title{CheckResponse Object}
\usage{
CheckResponse(operationId = NULL, checkErrors = NULL, checkInfo = NULL,
  quotaInfo = NULL, serviceConfigId = NULL)
}
\arguments{
\item{operationId}{The same operation_id value used in the CheckRequest}

\item{checkErrors}{Indicate the decision of the check}

\item{checkInfo}{Feedback data returned from the server during processing a Check request}

\item{quotaInfo}{Quota information for the check request associated with this response}

\item{serviceConfigId}{The actual config id used to process the request}
}
\value{
CheckResponse object
}
\description{
CheckResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Response message for the Check method.
}
