% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servicecontrol_objects.R
\name{EndReconciliationRequest}
\alias{EndReconciliationRequest}
\title{EndReconciliationRequest Object}
\usage{
EndReconciliationRequest(serviceConfigId = NULL,
  reconciliationOperation = NULL)
}
\arguments{
\item{serviceConfigId}{Specifies which version of service configuration should be used to process}

\item{reconciliationOperation}{Operation that describes the quota reconciliation}
}
\value{
EndReconciliationRequest object
}
\description{
EndReconciliationRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other EndReconciliationRequest functions: \code{\link{services.endReconciliation}}
}
