% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servicecontrol_objects.R
\name{EndReconciliationResponse}
\alias{EndReconciliationResponse}
\title{EndReconciliationResponse Object}
\usage{
EndReconciliationResponse(quotaMetrics = NULL, operationId = NULL,
  reconciliationErrors = NULL, serviceConfigId = NULL)
}
\arguments{
\item{quotaMetrics}{Metric values as tracked by One Platform before the adjustment was made}

\item{operationId}{The same operation_id value used in the EndReconciliationRequest}

\item{reconciliationErrors}{Indicates the decision of the reconciliation end}

\item{serviceConfigId}{ID of the actual config used to process the request}
}
\value{
EndReconciliationResponse object
}
\description{
EndReconciliationResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
