% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servicecontrol_objects.R
\name{ExplicitBuckets}
\alias{ExplicitBuckets}
\title{ExplicitBuckets Object}
\usage{
ExplicitBuckets(bounds = NULL)
}
\arguments{
\item{bounds}{'bound' is a list of strictly increasing boundaries between}
}
\value{
ExplicitBuckets object
}
\description{
ExplicitBuckets Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Describing buckets with arbitrary user-provided width.
}
