% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servicecontrol_objects.R
\name{LogEntry.labels}
\alias{LogEntry.labels}
\title{LogEntry.labels Object}
\usage{
LogEntry.labels()
}
\value{
LogEntry.labels object
}
\description{
LogEntry.labels Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A set of user-defined (key, value) data that provides additionalinformation about the log entry.
}
\seealso{
Other LogEntry functions: \code{\link{LogEntry.protoPayload}},
  \code{\link{LogEntry.structPayload}},
  \code{\link{LogEntry}}
}

