% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servicecontrol_objects.R
\name{LogEntry.protoPayload}
\alias{LogEntry.protoPayload}
\title{LogEntry.protoPayload Object}
\usage{
LogEntry.protoPayload()
}
\value{
LogEntry.protoPayload object
}
\description{
LogEntry.protoPayload Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The log entry payload, represented as a protocol buffer that isexpressed as a JSON object. You can only pass `protoPayload`values that belong to a set of approved types.
}
\seealso{
Other LogEntry functions: \code{\link{LogEntry.labels}},
  \code{\link{LogEntry.structPayload}},
  \code{\link{LogEntry}}
}
