% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servicecontrol_objects.R
\name{LogEntry.structPayload}
\alias{LogEntry.structPayload}
\title{LogEntry.structPayload Object}
\usage{
LogEntry.structPayload()
}
\value{
LogEntry.structPayload object
}
\description{
LogEntry.structPayload Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The log entry payload, represented as a structure thatis expressed as a JSON object.
}
\seealso{
Other LogEntry functions: \code{\link{LogEntry.labels}},
  \code{\link{LogEntry.protoPayload}},
  \code{\link{LogEntry}}
}
