% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servicecontrol_objects.R
\name{MetricValue.labels}
\alias{MetricValue.labels}
\title{MetricValue.labels Object}
\usage{
MetricValue.labels()
}
\value{
MetricValue.labels object
}
\description{
MetricValue.labels Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The labels describing the metric value.See comments on google.api.servicecontrol.v1.Operation.labels forthe overriding relationship.
}
\seealso{
Other MetricValue functions: \code{\link{MetricValue}}
}
