% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servicecontrol_objects.R
\name{MetricValue}
\alias{MetricValue}
\title{MetricValue Object}
\usage{
MetricValue(MetricValue.labels = NULL, boolValue = NULL, endTime = NULL,
  startTime = NULL, moneyValue = NULL, stringValue = NULL,
  labels = NULL, doubleValue = NULL, int64Value = NULL,
  distributionValue = NULL)
}
\arguments{
\item{MetricValue.labels}{The \link{MetricValue.labels} object or list of objects}

\item{boolValue}{A boolean value}

\item{endTime}{The end of the time period over which this metric value's measurement}

\item{startTime}{The start of the time period over which this metric value's measurement}

\item{moneyValue}{A money value}

\item{stringValue}{A text string value}

\item{labels}{The labels describing the metric value}

\item{doubleValue}{A double precision floating point value}

\item{int64Value}{A signed 64-bit integer value}

\item{distributionValue}{A distribution value}
}
\value{
MetricValue object
}
\description{
MetricValue Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Represents a single metric value.
}
\seealso{
Other MetricValue functions: \code{\link{MetricValue.labels}}
}
