% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servicecontrol_objects.R
\name{Operation.labels}
\alias{Operation.labels}
\title{Operation.labels Object}
\usage{
Operation.labels()
}
\value{
Operation.labels object
}
\description{
Operation.labels Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Labels describing the operation. Only the following labels are allowed:- Labels describing monitored resources as defined in  the service configuration.- Default labels of metric values. When specified, labels defined in the  metric value override these default.- The following labels defined by Google Cloud Platform:    - `cloud.googleapis.com/location` describing the location where the       operation happened,    - `servicecontrol.googleapis.com/user_agent` describing the user agent       of the API request,    - `servicecontrol.googleapis.com/service_agent` describing the service       used to handle the API request (e.g. ESP),    - `servicecontrol.googleapis.com/platform` describing the platform       where the API is served (e.g. GAE, GCE, GKE).
}
\seealso{
Other Operation functions: \code{\link{Operation}}
}

