% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servicecontrol_objects.R
\name{Operation}
\alias{Operation}
\title{Operation Object}
\usage{
Operation(Operation.labels = NULL, operationName = NULL,
  operationId = NULL, endTime = NULL, labels = NULL, importance = NULL,
  consumerId = NULL, startTime = NULL, logEntries = NULL,
  metricValueSets = NULL)
}
\arguments{
\item{Operation.labels}{The \link{Operation.labels} object or list of objects}

\item{operationName}{Fully qualified name of the operation}

\item{operationId}{Identity of the operation}

\item{endTime}{End time of the operation}

\item{labels}{Labels describing the operation}

\item{importance}{DO NOT USE}

\item{consumerId}{Identity of the consumer who is using the service}

\item{startTime}{Required}

\item{logEntries}{Represents information to be logged}

\item{metricValueSets}{Represents information about this operation}
}
\value{
Operation object
}
\description{
Operation Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Represents information regarding an operation.
}
\seealso{
Other Operation functions: \code{\link{Operation.labels}}
}

