% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servicecontrol_objects.R
\name{QuotaError}
\alias{QuotaError}
\title{QuotaError Object}
\usage{
QuotaError(subject = NULL, description = NULL, code = NULL)
}
\arguments{
\item{subject}{Subject to whom this error applies}

\item{description}{Free-form text that provides details on the cause of the error}

\item{code}{Error code}
}
\value{
QuotaError object
}
\description{
QuotaError Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
