% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servicecontrol_objects.R
\name{QuotaInfo.quotaConsumed}
\alias{QuotaInfo.quotaConsumed}
\title{QuotaInfo.quotaConsumed Object}
\usage{
QuotaInfo.quotaConsumed()
}
\value{
QuotaInfo.quotaConsumed object
}
\description{
QuotaInfo.quotaConsumed Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Map of quota group name to the actual number of tokens consumed. If thequota check was not successful, then this will not be populated due to noquota consumption.Deprecated: Use quota_metrics to get per quota group usage.
}
\seealso{
Other QuotaInfo functions: \code{\link{QuotaInfo}}
}
