% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servicecontrol_objects.R
\name{QuotaInfo}
\alias{QuotaInfo}
\title{QuotaInfo Object}
\usage{
QuotaInfo(QuotaInfo.quotaConsumed = NULL, quotaConsumed = NULL,
  quotaMetrics = NULL, limitExceeded = NULL)
}
\arguments{
\item{QuotaInfo.quotaConsumed}{The \link{QuotaInfo.quotaConsumed} object or list of objects}

\item{quotaConsumed}{Map of quota group name to the actual number of tokens consumed}

\item{quotaMetrics}{Quota metrics to indicate the usage}

\item{limitExceeded}{Quota Metrics that have exceeded quota limits}
}
\value{
QuotaInfo object
}
\description{
QuotaInfo Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Contains the quota information for a quota check response.
}
\seealso{
Other QuotaInfo functions: \code{\link{QuotaInfo.quotaConsumed}}
}
