% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servicecontrol_objects.R
\name{QuotaOperation.labels}
\alias{QuotaOperation.labels}
\title{QuotaOperation.labels Object}
\usage{
QuotaOperation.labels()
}
\value{
QuotaOperation.labels object
}
\description{
QuotaOperation.labels Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Labels describing the operation.
}
\seealso{
Other QuotaOperation functions: \code{\link{QuotaOperation}}
}
