% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servicecontrol_objects.R
\name{QuotaOperation}
\alias{QuotaOperation}
\title{QuotaOperation Object}
\usage{
QuotaOperation(QuotaOperation.labels = NULL, labels = NULL,
  consumerId = NULL, operationId = NULL, methodName = NULL,
  quotaMode = NULL, quotaMetrics = NULL)
}
\arguments{
\item{QuotaOperation.labels}{The \link{QuotaOperation.labels} object or list of objects}

\item{labels}{Labels describing the operation}

\item{consumerId}{Identity of the consumer for whom this quota operation is being performed}

\item{operationId}{Identity of the operation}

\item{methodName}{Fully qualified name of the API method for which this quota operation is}

\item{quotaMode}{Quota mode for this operation}

\item{quotaMetrics}{Represents information about this operation}
}
\value{
QuotaOperation object
}
\description{
QuotaOperation Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Represents information regarding a quota operation.
}
\seealso{
Other QuotaOperation functions: \code{\link{QuotaOperation.labels}}
}
