% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servicecontrol_objects.R
\name{QuotaProperties.limitByIds}
\alias{QuotaProperties.limitByIds}
\title{QuotaProperties.limitByIds Object}
\usage{
QuotaProperties.limitByIds()
}
\value{
QuotaProperties.limitByIds object
}
\description{
QuotaProperties.limitByIds Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
LimitType IDs that should be used for checking quota. Key in this mapshould be a valid LimitType string, and the value is the ID to be used. Forexample, an entry <USER, 123> will cause all user quota limits to use 123as the user ID. See google/api/quota.proto for the definition of LimitType.CLIENT_PROJECT: Not supported.USER: Value of this entry will be used for enforcing user-level quota      limits. If none specified, caller IP passed in the      servicecontrol.googleapis.com/caller_ip label will be used instead.      If the server cannot resolve a value for this LimitType, an error      will be thrown. No validation will be performed on this ID.Deprecated: use servicecontrol.googleapis.com/user label to send user ID.
}
\seealso{
Other QuotaProperties functions: \code{\link{QuotaProperties}}
}
