% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servicecontrol_objects.R
\name{QuotaProperties}
\alias{QuotaProperties}
\title{QuotaProperties Object}
\usage{
QuotaProperties(QuotaProperties.limitByIds = NULL, limitByIds = NULL,
  quotaMode = NULL)
}
\arguments{
\item{QuotaProperties.limitByIds}{The \link{QuotaProperties.limitByIds} object or list of objects}

\item{limitByIds}{LimitType IDs that should be used for checking quota}

\item{quotaMode}{Quota mode for this operation}
}
\value{
QuotaProperties object
}
\description{
QuotaProperties Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Represents the properties needed for quota operations.Use the metric_value_sets field in Operation message to provide costoverride with metric_name in <service_name>/quota/<quota_group_name>/costformat. Overrides for unmatched quota groups will be ignored.Costs are expected to be >= 0. Cost 0 will cause no quota check,but still traffic restrictions will be enforced.
}
\seealso{
Other QuotaProperties functions: \code{\link{QuotaProperties.limitByIds}}
}
