% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servicecontrol_objects.R
\name{ReleaseQuotaRequest}
\alias{ReleaseQuotaRequest}
\title{ReleaseQuotaRequest Object}
\usage{
ReleaseQuotaRequest(serviceConfigId = NULL, releaseOperation = NULL)
}
\arguments{
\item{serviceConfigId}{Specifies which version of service configuration should be used to process}

\item{releaseOperation}{Operation that describes the quota release}
}
\value{
ReleaseQuotaRequest object
}
\description{
ReleaseQuotaRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Request message for the ReleaseQuota method.
}
\seealso{
Other ReleaseQuotaRequest functions: \code{\link{services.releaseQuota}}
}
