% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servicecontrol_objects.R
\name{ReportError}
\alias{ReportError}
\title{ReportError Object}
\usage{
ReportError(operationId = NULL, status = NULL)
}
\arguments{
\item{operationId}{The Operation}

\item{status}{Details of the error when processing the `Operation`}
}
\value{
ReportError object
}
\description{
ReportError Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Represents the processing error of one `Operation` in the request.
}
