% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servicecontrol_objects.R
\name{ReportInfo}
\alias{ReportInfo}
\title{ReportInfo Object}
\usage{
ReportInfo(operationId = NULL, quotaInfo = NULL)
}
\arguments{
\item{operationId}{The Operation}

\item{quotaInfo}{Quota usage info when processing the `Operation`}
}
\value{
ReportInfo object
}
\description{
ReportInfo Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
