% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servicecontrol_objects.R
\name{ReportRequest}
\alias{ReportRequest}
\title{ReportRequest Object}
\usage{
ReportRequest(operations = NULL)
}
\arguments{
\item{operations}{Operations to be reported}
}
\value{
ReportRequest object
}
\description{
ReportRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Request message for the Report method.
}
\seealso{
Other ReportRequest functions: \code{\link{services.report}}
}

