% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servicecontrol_objects.R
\name{RequestMetadata}
\alias{RequestMetadata}
\title{RequestMetadata Object}
\usage{
RequestMetadata(callerIp = NULL, callerSuppliedUserAgent = NULL)
}
\arguments{
\item{callerIp}{The IP address of the caller}

\item{callerSuppliedUserAgent}{The user agent of the caller}
}
\value{
RequestMetadata object
}
\description{
RequestMetadata Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Metadata about the request.
}
