% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servicecontrol_functions.R
\docType{package}
\name{servicecontrol_googleAuthR}
\alias{servicecontrol_googleAuthR}
\alias{servicecontrol_googleAuthR-package}
\title{Google Service Control API
Google Service Control provides control plane functionality to managed services, such as logging, monitoring, and status checks.}
\description{
Auto-generated code by googleAuthR::gar_create_api_skeleton
 at 2016-09-03 23:28:55
filename: /Users/mark/dev/R/autoGoogleAPI/googleservicecontrolv1.auto/R/servicecontrol_functions.R
api_json: api_json
}
\details{
Authentication scopes used are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/servicecontrol
}
}

