% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servicecontrol_functions.R
\name{services.check}
\alias{services.check}
\title{Checks an operation with Google Service Control to decide whetherthe given operation should proceed. It should be called before theoperation is executed.This method requires the `servicemanagement.services.check` permissionon the specified service. For more information, see[Google Cloud IAM](https://cloud.google.com/iam).}
\usage{
services.check(CheckRequest, serviceName)
}
\arguments{
\item{CheckRequest}{The \link{CheckRequest} object to pass to this method}

\item{serviceName}{The service name as specified in its service configuration}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/servicecontrol
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform, https://www.googleapis.com/auth/servicecontrol)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/service-control/}{Google Documentation}

Other CheckRequest functions: \code{\link{CheckRequest}}
}

