% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servicecontrol_functions.R
\name{services.releaseQuota}
\alias{services.releaseQuota}
\title{Releases previously allocated quota done through AllocateQuota method.This method requires the `servicemanagement.services.quota`permission on the specified service. For more information, see[Google Cloud IAM](https://cloud.google.com/iam).}
\usage{
services.releaseQuota(ReleaseQuotaRequest, serviceName)
}
\arguments{
\item{ReleaseQuotaRequest}{The \link{ReleaseQuotaRequest} object to pass to this method}

\item{serviceName}{Name of the service as specified in the service configuration}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/servicecontrol
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform, https://www.googleapis.com/auth/servicecontrol)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/service-control/}{Google Documentation}

Other ReleaseQuotaRequest functions: \code{\link{ReleaseQuotaRequest}}
}
