% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servicecontrol_functions.R
\name{services.report}
\alias{services.report}
\title{Reports operation results to Google Service Control, such as logs andmetrics. It should be called after an operation is completed.If feasible, the client should aggregate reporting data for up to 5seconds to reduce API traffic. Limiting aggregation to 5 seconds is toreduce data loss during client crashes. Clients should carefully choosethe aggregation time window to avoid data loss risk more than 0.01%for business and compliance reasons.NOTE: the `ReportRequest` has the size limit of 1MB.This method requires the `servicemanagement.services.report` permissionon the specified service. For more information, see[Google Cloud IAM](https://cloud.google.com/iam).}
\usage{
services.report(ReportRequest, serviceName)
}
\arguments{
\item{ReportRequest}{The \link{ReportRequest} object to pass to this method}

\item{serviceName}{The service name as specified in its service configuration}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/servicecontrol
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform, https://www.googleapis.com/auth/servicecontrol)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/service-control/}{Google Documentation}

Other ReportRequest functions: \code{\link{ReportRequest}}
}
