% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servicecontrol_objects.R
\name{StartReconciliationRequest}
\alias{StartReconciliationRequest}
\title{StartReconciliationRequest Object}
\usage{
StartReconciliationRequest(reconciliationOperation = NULL,
  serviceConfigId = NULL)
}
\arguments{
\item{reconciliationOperation}{Operation that describes the quota reconciliation}

\item{serviceConfigId}{Specifies which version of service configuration should be used to process}
}
\value{
StartReconciliationRequest object
}
\description{
StartReconciliationRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other StartReconciliationRequest functions: \code{\link{services.startReconciliation}}
}
