% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servicecontrol_objects.R
\name{StartReconciliationResponse}
\alias{StartReconciliationResponse}
\title{StartReconciliationResponse Object}
\usage{
StartReconciliationResponse(quotaMetrics = NULL, operationId = NULL,
  reconciliationErrors = NULL, serviceConfigId = NULL)
}
\arguments{
\item{quotaMetrics}{Metric values as tracked by One Platform before the start of}

\item{operationId}{The same operation_id value used in the StartReconciliationRequest}

\item{reconciliationErrors}{Indicates the decision of the reconciliation start}

\item{serviceConfigId}{ID of the actual config used to process the request}
}
\value{
StartReconciliationResponse object
}
\description{
StartReconciliationResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
