% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servicecontrol_objects.R
\name{CheckError}
\alias{CheckError}
\title{Google Service Control API Objects 
Google Service Control provides control plane functionality to managed services, such as logging, monitoring, and status checks.}
\usage{
CheckError(code = NULL, detail = NULL)
}
\arguments{
\item{code}{The error code}

\item{detail}{Free-form text providing details on the error cause of the error}
}
\value{
CheckError object
}
\description{
Auto-generated code by googleAuthR::gar_create_api_objects
 at 2016-09-03 23:28:55
filename: /Users/mark/dev/R/autoGoogleAPI/googleservicecontrolv1.auto/R/servicecontrol_objects.R
api_json: api_json
}
\details{
Objects for use by the functions created by googleAuthR::gar_create_api_skeleton
CheckError Object


Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Defines the errors to be returned ingoogle.api.servicecontrol.v1.CheckResponse.check_errors.
}

