% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servicecontrol_objects.R
\name{CheckRequest}
\alias{CheckRequest}
\title{CheckRequest Object}
\usage{
CheckRequest(operation = NULL)
}
\arguments{
\item{operation}{The operation to be checked}
}
\value{
CheckRequest object
}
\description{
CheckRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Request message for the Check method.
}
\seealso{
Other CheckRequest functions: \code{\link{services.check}}
}

