% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servicecontrol_objects.R
\name{CheckResponse}
\alias{CheckResponse}
\title{CheckResponse Object}
\usage{
CheckResponse(checkErrors = NULL, operationId = NULL,
  serviceConfigId = NULL)
}
\arguments{
\item{checkErrors}{Indicate the decision of the check}

\item{operationId}{The same operation_id value used in the CheckRequest}

\item{serviceConfigId}{The actual config id used to process the request}
}
\value{
CheckResponse object
}
\description{
CheckResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Response message for the Check method.
}

