% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servicecontrol_objects.R
\name{Distribution}
\alias{Distribution}
\title{Distribution Object}
\usage{
Distribution(explicitBuckets = NULL, maximum = NULL,
  sumOfSquaredDeviation = NULL, exponentialBuckets = NULL, minimum = NULL,
  linearBuckets = NULL, mean = NULL, count = NULL, bucketCounts = NULL)
}
\arguments{
\item{explicitBuckets}{Buckets with arbitrary user-provided width}

\item{maximum}{The maximum of the population of values}

\item{sumOfSquaredDeviation}{The sum of squared deviations from the mean:}

\item{exponentialBuckets}{Buckets with exponentially growing width}

\item{minimum}{The minimum of the population of values}

\item{linearBuckets}{Buckets with constant width}

\item{mean}{The arithmetic mean of the samples in the distribution}

\item{count}{The total number of samples in the distribution}

\item{bucketCounts}{The number of samples in each histogram bucket}
}
\value{
Distribution object
}
\description{
Distribution Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Distribution represents a frequency distribution of double-valued samplepoints. It contains the size of the population of sample points plusadditional optional information:  - the arithmetic mean of the samples  - the minimum and maximum of the samples  - the sum-squared-deviation of the samples, used to compute variance  - a histogram of the values of the sample points
}
