% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servicecontrol_objects.R
\name{ExponentialBuckets}
\alias{ExponentialBuckets}
\title{ExponentialBuckets Object}
\usage{
ExponentialBuckets(numFiniteBuckets = NULL, growthFactor = NULL,
  scale = NULL)
}
\arguments{
\item{numFiniteBuckets}{The number of finite buckets}

\item{growthFactor}{The i'th exponential bucket covers the interval}

\item{scale}{The i'th exponential bucket covers the interval}
}
\value{
ExponentialBuckets object
}
\description{
ExponentialBuckets Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Describing buckets with exponentially growing width.
}
