% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servicecontrol_objects.R
\name{LinearBuckets}
\alias{LinearBuckets}
\title{LinearBuckets Object}
\usage{
LinearBuckets(offset = NULL, numFiniteBuckets = NULL, width = NULL)
}
\arguments{
\item{offset}{The i'th linear bucket covers the interval}

\item{numFiniteBuckets}{The number of finite buckets}

\item{width}{The i'th linear bucket covers the interval}
}
\value{
LinearBuckets object
}
\description{
LinearBuckets Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Describing buckets with constant width.
}
