% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servicecontrol_objects.R
\name{LogEntry}
\alias{LogEntry}
\title{LogEntry Object}
\usage{
LogEntry(LogEntry.structPayload = NULL, LogEntry.labels = NULL,
  LogEntry.protoPayload = NULL, textPayload = NULL, structPayload = NULL,
  labels = NULL, severity = NULL, name = NULL, protoPayload = NULL,
  timestamp = NULL, insertId = NULL)
}
\arguments{
\item{LogEntry.structPayload}{The \link{LogEntry.structPayload} object or list of objects}

\item{LogEntry.labels}{The \link{LogEntry.labels} object or list of objects}

\item{LogEntry.protoPayload}{The \link{LogEntry.protoPayload} object or list of objects}

\item{textPayload}{The log entry payload, represented as a Unicode string (UTF-8)}

\item{structPayload}{The log entry payload, represented as a structure that}

\item{labels}{A set of user-defined (key, value) data that provides additional}

\item{severity}{The severity of the log entry}

\item{name}{Required}

\item{protoPayload}{The log entry payload, represented as a protocol buffer that is}

\item{timestamp}{The time the event described by the log entry occurred}

\item{insertId}{A unique ID for the log entry used for deduplication}
}
\value{
LogEntry object
}
\description{
LogEntry Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
An individual log entry.
}
\seealso{
Other LogEntry functions: \code{\link{LogEntry.labels}},
  \code{\link{LogEntry.protoPayload}},
  \code{\link{LogEntry.structPayload}}
}

