% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servicecontrol_objects.R
\name{MetricValueSet}
\alias{MetricValueSet}
\title{MetricValueSet Object}
\usage{
MetricValueSet(metricName = NULL, metricValues = NULL)
}
\arguments{
\item{metricName}{The metric name defined in the service configuration}

\item{metricValues}{The values in this metric}
}
\value{
MetricValueSet object
}
\description{
MetricValueSet Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Represents a set of metric values in the same metric.Each metric value in the set should have a unique combination of start time,end time, and label values.
}

