% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servicecontrol_objects.R
\name{Operation}
\alias{Operation}
\title{Operation Object}
\usage{
Operation(Operation.labels = NULL, Operation.userLabels = NULL,
  quotaProperties = NULL, consumerId = NULL, operationId = NULL,
  endTime = NULL, operationName = NULL, startTime = NULL,
  importance = NULL, resourceContainer = NULL, labels = NULL,
  logEntries = NULL, userLabels = NULL, metricValueSets = NULL)
}
\arguments{
\item{Operation.labels}{The \link{Operation.labels} object or list of objects}

\item{Operation.userLabels}{The \link{Operation.userLabels} object or list of objects}

\item{quotaProperties}{Represents the properties needed for quota check}

\item{consumerId}{Identity of the consumer who is using the service}

\item{operationId}{Identity of the operation}

\item{endTime}{End time of the operation}

\item{operationName}{Fully qualified name of the operation}

\item{startTime}{Required}

\item{importance}{DO NOT USE}

\item{resourceContainer}{The resource name of the parent of a resource in the resource hierarchy}

\item{labels}{Labels describing the operation}

\item{logEntries}{Represents information to be logged}

\item{userLabels}{User defined labels for the resource that this operation is associated}

\item{metricValueSets}{Represents information about this operation}
}
\value{
Operation object
}
\description{
Operation Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Represents information regarding an operation.
}
\seealso{
Other Operation functions: \code{\link{Operation.labels}},
  \code{\link{Operation.userLabels}}
}
