% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servicecontrol_objects.R
\name{ReportResponse}
\alias{ReportResponse}
\title{ReportResponse Object}
\usage{
ReportResponse(reportErrors = NULL, serviceConfigId = NULL)
}
\arguments{
\item{reportErrors}{Partial failures, one for each `Operation` in the request that failed}

\item{serviceConfigId}{The actual config id used to process the request}
}
\value{
ReportResponse object
}
\description{
ReportResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Response message for the Report method.
}

